apis/iot/requirements.txt
gunicorn==21.2.0
fastapi==0.103.1
uvicorn==0.23.2
mysql-connector-python